recipe/src/main/java/org/libermundi/recipe/converters/CategoryCommandToCategory.java
package org.libermundi.recipe.converters;

import lombok.Synchronized;
import org.libermundi.recipe.commands.CategoryCommand;
import org.libermundi.recipe.domain.Category;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;

public class CategoryCommandToCategory implements Converter<CategoryCommand, Category> {
    @Nullable
    @Override
    @Synchronized
    public Category convert(CategoryCommand categoryCommand) {
        if(categoryCommand == null) {
            return null;
        }

        final Category category = new Category();

        category.setId(categoryCommand.getId());
        category.setName(categoryCommand.getName());

        return category;
    }
}
